Add a fallback route to index.html in development
Ember apps are single-page. Ember appkit requires that you enter the app
via index.html or another pre-defined file in public. If you have
dynamic URLs, this sucks. You can't just reload a page and be back at
the state you were in a moment before.

This adds wildcard middleware to connect. If a file/directory cannot be
loaded via the traditional pipeline, then the wildcard picks it up and
tries to return index.html or another file configured as the `wildcard`
property on the connect options.

So reload that dynamic URL at will, ye masters of Ember.