Naive source-maps with grunt-concat-sourcemap
These source maps are generated at concat time from the transpiled
output, so they of course don't map 1:1 with the original lines. Still,
it's a start, and keeps devs from having to debug a giant concatenated
file. (the transpiler output is reasonably readable, of course)

Also dropped the camelize in the Gruntfile's config loader, as the task
name for grunt-concat-sourcemap is actually `concat_sourcemap`.