refactor api-stub
Refactor the api-stub to support dynamic segments.

This is a little verbose, but is a does allow for actual dynamic segment ids to be requested from the api stub, as well as collections.

A fake id is used from the request param is no matching id is found.

Collection
http://0.0.0.0:8000/api/posts/

Individual record
http://0.0.0.0:8000/api/posts/1

Arbitrary record id
http://0.0.0.0:8000/api/posts/12345

Added colors model
http://0.0.0.0:8000/api/colors/
http://0.0.0.0:8000/api/colors/1
http://0.0.0.0:8000/api/colors/2
etc...