Adds test helper for asserting a given route is active
This helper is useful for verifying route transitions. After a visit()
or click(), you can use this helper to verify the end state of a
redirect or transitionTo().

Usage: currentRoute( 'posts.new', 'posts/new route is active' )

It uses the isolated container to get the main router and, from that,
the current active handler. Given a route name, it asserts that the
provided name is equal to the route name of the active handler.