[#401] Fakes testem.js so that in-browser tests run
Testem requires a script tag to load /testem.js, which is supplied
by Testem. This file doesn't actually exist on the file system, which
causes localhost:8000/tests to fail with a global failure.

By telling express to serve up a fake testem.js file, the script load
is satisfied, thus making it posible to execute the tests in both
the Testem and direct contexts.