Fixes URL routing
The way the static middleware works, to serve a file but also let
index.html catch everything, the urlRoot option must be used in
conjunction with the directory option.

Previously, the introduction of faking the testem.js file broke
location history routing, because all routes were caught by the
testem fake file. This patch serves the fake file for the URL, but
also preserves the index.html catch all.