Allow clearing the registry.
This is useful for testing custom lookups in the resolver itself. I
would much prefer to keep this the same everywhere and it currently
seems that the canonical `loader.js` is here in EAK.

Also, moved the definition of `requirejs._eak_seen` to be outside of the
`requirejs` function. I am not sure why we would want to reset it
everytime that we call require.