Don't slice if parentBase has just one element.
require('ember-qunit') was failing because we couldn't  find module './isolated-container'.

ember-qunit allow us to do require('ember-qunit') which internally does "import isolatedContainer from './isolated-container"

When requiring, the loader was calling "resolve" with
child="./isolatedContainer" and name = "ember-qunit", resolving to
'./isolated-container' which is wrong, it should have been
'ember-qunit/isolated-container', the reason is that

 "ember-qunit".split('/').slice(0, -1) === []

Is this is a case we didn't handle or are parents names suppose to
have "/"?