Don't use PORT environment variable for grunt-express-server;
Setting the environment variables for port configurations feels a bit odd and is obsecure to locate. In addition, `PORT` is a very commons environment variable that may collide with others.

By defining it explicitly in package.json, the user is able to quickly and easily change the port and check it into source control.